encapsulated csv read and write logic into seperate class
